+++ ESFramework/Core/Additions/NSUserDefaults+ESAdditions.h
@@ -10,6 +10,24 @@
 
 @interface NSUserDefaults (ESAdditions)
 
+/**
+ * Set an object for the property identified by a given key path to a given value.
+ *
+ * @param    object                  The object for the property identified by _keyPath_.
+ * @param    keyPath                 A key path of the form _relationship.property_ (with one or more relationships): for example “department.name” or “department.manager.lastName.”
+ */
+- (void)setObject:(id)object forKeyPath:(NSString *)keyPath;
+
+/**
+ * Set an object for the property identified by a given key path to a given value, with optional parameters to control creation and replacement of intermediate objects.
+ *
+ * @param    object                  The object for the property identified by _keyPath_.
+ * @param    keyPath                 A key path of the form _relationship.property_ (with one or more relationships): for example “department.name” or “department.manager.lastName.”
+ * @param    createIntermediates     Intermediate dictionaries defined within the key path that do not currently exist in the receiver are created.
+ * @param    replaceIntermediates    Intermediate objects encountered in the key path that are not a direct subclass of `NSDictionary` are replaced.
+ */
+- (void)setObject:(id)object forKeyPath:(NSString *)keyPath createIntermediateDictionaries:(BOOL)createIntermediates replaceIntermediateObjects:(BOOL)replaceIntermediates;
+
 /// =============================================
 /// @name Quick methods for standardUserDefaults
 /// =============================================